Attempt to further reduce test flakiness
This plugin suffers from a large amount of test flakiness in the travis
ci environment. This commit continues the attempts to reduce this flakiness,
starting with https://github.com/logstash-plugins/logstash-input-file/pull/263

Mostly, this is increasing timeouts to deal with sluggish build servers, but also
includes a fix to a Stud.try call which would not retry due to failure exceptions
not matching.