Fixes asynch test on autoflush
With the fix on multiline codec (https://github.com/logstash-plugins/logstash-codec-multiline/pull/70) introduced
to avoid sharing of codec across multiple threads, it started to make fail this test that leveraged that shared instance
to tracing the codec calls.
The multithreading nature of the test and the lack of synchornizaion to access the IdentityCodecMap led to multiple threads
overwrite of the codec instances. This overwrite cleans the history of previous tracked calls making it failing.

This commit fixes it giving to che #clone a method a simil-singleton functionality, to don't loose the list of tracked calls.