Doc: Set version attribute in plugin source file (#308)
Update links to long format to pick up version
Bump to v4.4.1

Note that we're resetting the version attribute to `current` at the end of the file to avoid carry over to other files. After we resolve formatting and clean up older files, we can remove the reset in the source file.